Fix up Cabaon (Part 1)
* Tweaked the map slightly
* Changed some control of coastal Tripolitana

Cabaon is a person and besides the map from the 1960s, there's no evidences of any permanence of control in Tripoli instead of a victory over the Vandals. Additionally, the character Cabaon was not a Christian, but a Berber and already is part of the Lagautan tribe.

Part 1 excises Cabaon, fixes up the barony alignment, and moves control around to a more accurate model of Late Antique Africa.